Part 4.21 complete

+++ part4/Blogilista/controllers/blogs.js
@@ -45,6 +45,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
+  if (!decodedToken.id)
+    return response.status(401).json({ error: 'token invalid' })
+
+  const blog = await Blog.findById(request.params.id)
+  console.log(blog.user.toString());
+  console.log(decodedToken.id.toString());
+
+  if (!(blog.user.toString() === decodedToken.id.toString()))
+    return response.status(401).json({ error: 'can delete only own items' })
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
